toolkit/components/passwordmgr/test/mochitest/test_autocomplete_new_password.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test autofill and autocomplete on autocomplete=new-password fields</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <script src="pwmgr_common.js"></script>
  <script src="../../../satchel/test/satchel_common.js"></script>
  <link rel="stylesheet" href="/tests/SimpleTest/test.css" />
</head>
<body>
Login Manager test: autofill with autocomplete=new-password fields

<script>
function initLogins() {
  const {Services} = ChromeUtils.import("resource://gre/modules/Services.jsm");
  Services.logins.removeAllLogins();
  let login1 = Cc["@mozilla.org/login-manager/loginInfo;1"].createInstance(Ci.nsILoginInfo);
  login1.init("https://example.com", "https://autofill", null, "user1", "pass1");

  const {LoginManagerParent} = ChromeUtils.import("resource://gre/modules/LoginManagerParent.jsm");
  Services.telemetry.clearEvents();
  if (LoginManagerParent._generatedPasswordsByPrincipalOrigin) {
    LoginManagerParent._generatedPasswordsByPrincipalOrigin.clear();
  }
  Services.logins.addLogin(login1);
}

runInParent(initLogins);

let readyPromise = registerRunTests();
</script>
<p id="display"></p>

<!-- we presumably can't hide the content for this test. -->
<div id="content">

  <!-- form1 is the reference, sanity-check -->
  <form id="form1" action="https://autofill" onsubmit="return false;">
    <input type="text" name="uname">
    <input type="password" name="p">
    <button type="submit">Submit</button>
  </form>

  <!-- form2 uses a new-password type=password field -->
  <form id="form2" action="https://autofill" onsubmit="return false;">
    <input type="text" name="uname">
    <input type="password" name="pword" autocomplete="new-password">
    <button type="submit">Submit</button>
  </form>
</div>

<pre id="test">
<script class="testbody" type="text/javascript">
const {ContentTaskUtils} =
  SpecialPowers.Cu.import("resource://testing-common/ContentTaskUtils.jsm", {});
const { TestUtils } = SpecialPowers.Cu.import("resource://testing-common/TestUtils.jsm");

let dateAndTimeFormatter = new SpecialPowers.Services.intl.DateTimeFormat(undefined, {
  dateStyle: "medium",
});

const TelemetryFilterProps = Object.freeze({
  category: "pwmgr",
  method: "autocomplete_field",
  object: "generatedpassword",
});

async function waitForTelemetryEventsCondition(cond, options = {},
                                               errorMsg = "waitForTelemetryEventsCondition timed out", maxTries = 200) {
  return TestUtils.waitForCondition(async () => {
    let events = await getTelemetryEvents(options);
    let result;
    try {
      result = cond(events);
      info("waitForTelemetryEventsCondition, result: " + result);
    } catch (e) {
      info("waitForTelemetryEventsCondition caught exception, got events: " + JSON.stringify(events));
      ok(false, `${e}\n${e.stack}`);
    }
    return result ? events : null;
  }, errorMsg, maxTries);
}

add_task(async function setup() {
  ok(readyPromise, "check promise is available");
  await readyPromise;
  let events = await getTelemetryEvents({ process: "parent", filterProps: TelemetryFilterProps, clear: true });
  is(events.length, 0, "Expect 0 events");
});

add_task(async function test_autofillAutocompleteUsername_noGeneration() {
  // reference form was filled as expected?
  checkForm(1, "user1", "pass1");

  // 2nd form should not be filled
  checkForm(2, "", "");

  $_(2, "uname").focus();
  const shownPromise = promiseACShown();
  synthesizeKey("KEY_ArrowDown");
  let results = await shownPromise;
  let expectedACLabels = ["user1"];
  checkAutoCompleteResults(results, expectedACLabels, "example.com", "Check all rows are correct");
  synthesizeKey("KEY_ArrowDown");
  synthesizeKey("KEY_Enter");
  await promiseFormsProcessed();
  checkForm(2, "user1", "pass1");

  document.getElementById("form2").reset();
});

add_task(async function test_autofillAutocompletePassword_noGeneration() {
  // 2nd form should not be filled
  checkForm(2, "", "");

  let pword = $_(2, "pword");
  pword.focus();
  const shownPromise = promiseACShown();
  synthesizeKey("KEY_ArrowDown");
  let results = await shownPromise;
  let expectedACLabels = ["user1"];
  checkAutoCompleteResults(results, expectedACLabels, "example.com", "Check all rows are correct");
  synthesizeKey("KEY_ArrowDown");
  synthesizeKey("KEY_Enter");
  // Can't use promiseFormsProcessed() when autocomplete fills the field directly.
  await SimpleTest.promiseWaitForCondition(() => pword.value == "pass1", "Check pw filled");
  checkForm(2, "", "pass1");

  // No autocomplete results should appear for non-empty pw fields.
  synthesizeKey("KEY_ArrowDown");
  await promiseNoUnexpectedPopupShown();

  document.getElementById("form2").reset();
});

// All tests below this are with generation prefs enabled.

add_task(async function test_autofillAutocompleteUsername_noGeneration() {
  await SpecialPowers.pushPrefEnv({"set": [
    ["signon.generation.available", true],
    ["signon.generation.enabled", true],
    ["signon.showAutoCompleteOrigins", true],
  ]});

  // 2nd form should not be filled
  checkForm(2, "", "");

  $_(2, "uname").focus();
  const shownPromise = promiseACShown();
  synthesizeKey("KEY_ArrowDown");
  let results = await shownPromise;
  // No generation option on username fields.
  let expectedACLabels = ["user1"];
  checkAutoCompleteResults(results, expectedACLabels, "example.com", "Check all rows are correct");
  synthesizeKey("KEY_ArrowDown");
  synthesizeKey("KEY_Enter");
  await promiseFormsProcessed();
  checkForm(2, "user1", "pass1");

  document.getElementById("form2").reset();
});

add_task(async function test_autofillAutocompletePassword_withGeneration() {
  // 2nd form should not be filled
  checkForm(2, "", "");

  let pword = $_(2, "pword");
  pword.focus();
  let shownPromise = promiseACShown();
  synthesizeKey("KEY_ArrowDown");
  let results = await shownPromise;
  let expectedACLabels = [
    "user1",
    "Use Generated Password",
  ];
  checkAutoCompleteResults(results, expectedACLabels, "example.com", "Check all rows are correct");
  synthesizeKey("KEY_ArrowDown");
  synthesizeKey("KEY_Enter");
  // Can't use promiseFormsProcessed() when autocomplete fills the field directly.
  await SimpleTest.promiseWaitForCondition(() => pword.value == "pass1", "Check pw filled");
  checkForm(2, "", "pass1");

  // No autocomplete results should appear for non-empty pw fields.
  synthesizeKey("KEY_ArrowDown");
  await promiseNoUnexpectedPopupShown();

  info("Removing all logins to test auto-saving of generated passwords");
  await LoginManager.removeAllLogins();

  while (pword.value) {
    synthesizeKey("KEY_Backspace");
  }

  info("This time select the generated password");
  shownPromise = promiseACShown();
  synthesizeKey("KEY_ArrowDown");
  results = await shownPromise;
  expectedACLabels = [
    "Use Generated Password",
  ];
  checkAutoCompleteResults(results, expectedACLabels, "example.com", "Check all rows are correct");
  synthesizeKey("KEY_ArrowDown");
  let storageAddPromise = promiseStorageChanged(["addLogin"]);
  synthesizeKey("KEY_Enter");

  info("waiting for the password field to be filled with the generatedpassword");
  await SimpleTest.promiseWaitForCondition(() => !!pword.value, "Check generated pw filled");
  info("Wait for generated password to be added to storage");
  await storageAddPromise;

  let expectedCount = 1;
  info(`filled generated password, check there are now ${expectedCount} generatedpassword telemetry events`);
  await waitForTelemetryEventsCondition(events => {
    return events.length == expectedCount;
  }, { process: "parent", filterProps: TelemetryFilterProps }, `Wait for there to be ${expectedCount} telemetry events`);

  let logins = await LoginManager.getAllLogins();
  let timePasswordChanged = logins[logins.length - 1].timePasswordChanged;
  let time = dateAndTimeFormatter.format(new Date(timePasswordChanged));
  const LABEL_NO_USERNAME = "No username (" + time + ")";

  let generatedPW = pword.value;
  is(generatedPW.length, GENERATED_PASSWORD_LENGTH, "Check generated password length");
  ok(generatedPW.match(GENERATED_PASSWORD_REGEX), "Check generated password format");

  while (pword.value) {
    synthesizeKey("KEY_Backspace");
  }

  shownPromise = promiseACShown();
  synthesizeKey("KEY_ArrowDown");
  results = await shownPromise;
  expectedACLabels = [
    LABEL_NO_USERNAME,
    "Use Generated Password",
  ];
  checkAutoCompleteResults(results, expectedACLabels, "example.com", "Check all rows are correct");
  synthesizeKey("KEY_ArrowDown");
  synthesizeKey("KEY_ArrowDown");
  synthesizeKey("KEY_Enter");
  await SimpleTest.promiseWaitForCondition(() => !!pword.value, "Check generated pw filled");
  // Same generated password should be used.
  checkForm(2, "", generatedPW);

  info("filled generated password again, ensure we don't record another generatedpassword autocomplete telemetry event");
  let telemetryEvents;
  expectedCount = 2;
  try {
    telemetryEvents = await waitForTelemetryEventsCondition(events => events.length == expectedCount,
                                                            { process: "parent", filterProps: TelemetryFilterProps },
                                                            `Wait for there to be ${expectedCount} events`, 50);
  } catch (ex) {}
  ok(!telemetryEvents, "Expected to timeout waiting for there to be 2 events");

  logins = await LoginManager.getAllLogins();
  is(logins.length, 1, "Still 1 login after filling the generated password a 2nd time");
  is(logins[0].timePasswordChanged, timePasswordChanged, "Saved login wasn't changed");
  is(logins[0].password, generatedPW, "Password is the same");

  info("reset initial login state with login1");
  runInParent(initLogins);
  info("invalidate the autocomplete cache after updating storage above");
  synthesizeKey("KEY_Backspace");
  document.getElementById("form2").reset();
});

add_task(async function test_autofillAutocompletePassword_saveLoginDisabled() {
  // form should not be filled
  checkForm(2, "", "");
  let formOrigin = new URL(document.documentURI).origin;
  is(formOrigin, "https://example.com", "Expected form origin");

  await LoginManager.setLoginSavingEnabled("https://example.com", false);

  let pword = $_(2, "pword");
  pword.focus();
  let shownPromise = promiseACShown();
  synthesizeKey("KEY_ArrowDown");
  let results = await shownPromise;
  let expectedACLabels = [
    // when login-saving is disabled for an origin, we expect no generated password row here
    "user1",
  ];
  info("ac results: " + results.join(", "));
  checkAutoCompleteResults(results, expectedACLabels, "example.com", "Check all rows are correct");

  document.getElementById("form2").reset();
  await LoginManager.setLoginSavingEnabled("https://example.com", true);
});

</script>
</pre>
</body>
</html>
